Refactor dashboard and utilities
