+++ src/components/ExternalControlInfo.tsx
